// web-application/src/types-openapi/api/models/ListCourtRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListCourtRequest
 */
export interface ListCourtRequest {
    /**
     * 
     * @type {string}
     * @memberof ListCourtRequest
     */
    name?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ListCourtRequest
     */
    courtAreaId?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ListCourtRequest
     */
    status?: string | null;
}

/**
 * Check if a given object implements the ListCourtRequest interface.
 */
export function instanceOfListCourtRequest(value: object): value is ListCourtRequest {
    return true;
}

export function ListCourtRequestFromJSON(json: any): ListCourtRequest {
    return ListCourtRequestFromJSONTyped(json, false);
}

export function ListCourtRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListCourtRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'courtAreaId': json['courtAreaId'] == null ? undefined : json['courtAreaId'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function ListCourtRequestToJSON(json: any): ListCourtRequest {
    return ListCourtRequestToJSONTyped(json, false);
}

export function ListCourtRequestToJSONTyped(value?: ListCourtRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'courtAreaId': value['courtAreaId'],
        'status': value['status'],
    };
}